Fix Token List on vuejs The component created was not working it was only looking ok as jquery writed over the html
