+++ src/components/crossForm/CrossForm.vue
@@ -125,7 +125,7 @@
         </div>
 
         <!-- Column 8 -->
-        <div class="receiveAddress text-center col-sm-5">
+        <div class="receiverAddress text-center col-sm-5">
           <label class="receive-label-address" for="receive-address"
             ><a id="same-address" class="same-address">Receiver address</a></label
           >
