--- src/constants/index.js
@@ -1,2 +1,0 @@
-export * from './networks'
-export * from './tokens'
